fix create negative address

+++ demo/src/test/java/global/eska/ddk/keygen/demo/DemoApplicationTests.java
@@ -11,13 +11,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
+import java.math.BigInteger;
 import java.util.Arrays;
 
 import static global.eska.ddk.keygen.utils.Etalon.*;
 import static global.eska.ddk.keygen.utils.HashUtils.sha256;
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.*;
 
 @Slf4j
 @RunWith(SpringRunner.class)
@@ -54,7 +53,7 @@ public class DemoApplicationTests {
         assertArrayEquals(PRIVATE_KEY_BYTES, keyPair.getSecretKey());
         assertEquals("PUBLIC_KEY_HEX", PUBLIC_KEY_HEX, keyPair.getPublicKeyHex());
 
-        long address = accountCreator.getAddressByPublicKey(keyPair.getPublicKey());
+        BigInteger address = accountCreator.getAddressByPublicKey(keyPair.getPublicKey());
         assertEquals("ADDRESS", ADDRESS, address);
 
         log.info("{}", PASSPHRASE);
